Update readme.md with additional training material
I added a blog resource to the list that could be valuable to learners interested in learning HAL topics. I personally found it hard to navigate the different HALs when starting out based on the existing resources. I created the added blog in the hope to help newcomers that could be in the same position I was in. It's a blog I update on weekly basis with a new topic relevant to embedded Rust.  This is the link https://apollolabsblog.hashnode.dev/series/stm32f4-embedded-rust-hal 

On a side note, I had attempted to submit a pull request before though I think I did something wrong where it didn't make it through somehow.